# app-principal/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Raiz Conecta API",
    "version": "1.0.0",
    "description": "API da plataforma Raiz Conecta — conecta produtores rurais e mercados. Gerencia autenticação, perfis, demandas, ofertas e administração de usuários."
  },
  "servers": [
    {
      "url": "https://seu-app.vercel.app",
      "description": "Produção (Vercel)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Desenvolvimento local"
    }
  ],
...
  "paths": {
...
    "/api/perfil/completar": {
      "post": {
        "tags": [
          "Perfil"
        ],
        "summary": "Completar perfil após cadastro",
        "description": "Recebe dados complementares (documento, endereço) e um arquivo de documento via `multipart/form-data`. Coloca o usuário em status `EM_ANALISE` aguardando aprovação do admin.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CompletarPerfilInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Perfil atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MensagemSucesso"
                },
                "example": {
                  "message": "Perfil atualizado com sucesso!"
                }
              }
            }
          },
          "400": {
            "description": "Dados obrigatórios ausentes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                },
                "example": {
                  "error": "Dados obrigatórios incompletos"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ErroInterno"
          }
        }
      }
    },
...
    "/api/vendedor/perfil": {
      "get": {
        "tags": [
          "Produtor"
        ],
        "summary": "Buscar perfil do produtor",
        "description": "Retorna os dados do produtor (vendedor) pelo e-mail.",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "email"
            },
            "example": "fazenda@email.com"
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do produtor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vendedor"
                }
              }
            }
          },
          "400": {
            "description": "E-mail não fornecido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                },
                "example": {
                  "error": "E-mail não fornecido"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ErroInterno"
          }
        }
      },
      "post": {
        "tags": [
          "Produtor"
        ],
        "summary": "Upload de documento do produtor",
        "description": "Recebe um arquivo de documento via `multipart/form-data`, salva em `/uploads/docs/` e atualiza o status do produtor para `A_ANALISE`.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "file"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Arquivo do documento (PDF, imagem, etc.)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Documento enviado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MensagemSucesso"
                },
                "example": {
                  "message": "Documento enviado com sucesso!"
                }
              }
            }
          },
          "400": {
            "description": "Dados ou arquivo faltando",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                },
                "example": {
                  "error": "Dados ou arquivo faltando"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ErroInterno"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "MensagemSucesso": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "Erro": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "CompletarPerfilInput": {
        "type": "object",
        "required": [
          "email",
          "tipoUsuario",
          "documento",
          "cep"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "tipoUsuario": {
            "type": "string",
            "enum": [
              "produtor",
              "mercado"
            ]
          },
          "tipoDoc": {
            "type": "string",
            "description": "Tipo do documento (CPF, CNPJ, etc.)"
          },
          "documento": {
            "type": "string"
          },
          "cep": {
            "type": "string"
          },
          "rua": {
            "type": "string"
          },
          "numero": {
            "type": "string"
          },
          "bairro": {
            "type": "string"
          },
          "cidade": {
            "type": "string"
          },
          "estado": {
            "type": "string"
          },
          "file": {
            "type": "string",
            "format": "binary",
            "description": "Arquivo do documento comprobatório"
          }
        }
      },
...
      "Vendedor": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "nomeFantasia": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "EM_ANALISE",
              "APROVADO",
              "REJEITADO",
              "A_ANALISE",
              "SUSPENSO"
            ]
          },
          "urlDocumento": {
            "type": "string",
            "nullable": true
          },
          "telefone": {
            "type": "string",
            "nullable": true
          },
          "cep": {
            "type": "string",
            "nullable": true
          },
          "cidade": {
            "type": "string",
            "nullable": true
          },
          "estado": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "responses": {
      "ErroInterno": {
        "description": "Erro interno no servidor",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Erro"
            },
            "example": {
              "error": "Erro interno no servidor"
            }
          }
        }
      }
    }
  }
}